+++ physics/src/include/IceMinima.hpp
@@ -5,6 +5,8 @@
 #ifndef ICEMINIMA_HPP
 #define ICEMINIMA_HPP
 
+#include "include/FloatType.hpp"
+
 namespace Nextsim {
 
 class IColumnPhysics;
@@ -12,14 +14,14 @@ class IColumnPhysics;
 //! IColumnPhysics and its dependencies.
 class IceMinima {
 public:
-    static inline double h() { return hMin; };
-    static inline double c() { return cMin; };
-    static const double hMinDefault;
-    static const double cMinDefault;
+    static inline FloatType h() { return hMin; };
+    static inline FloatType c() { return cMin; };
+    static const FloatType hMinDefault;
+    static const FloatType cMinDefault;
 
 private:
-    static double hMin;
-    static double cMin;
+    static FloatType hMin;
+    static FloatType cMin;
 
     friend IColumnPhysics;
 };
